Exposed RequiresAuth for plugins
